Make the size of the squares in the grid change depending on the rotation of the left controler
For this I first used a function to rotate an image, in this case the left controler.
Using the rotation value from that function I created an array of diferent posbile square sizes
which were all integers and divisors of the total size of the grid so that no matter the rotation
of the controler, the grid would be divided evenly.

To chose the value based on the rotation value which was a floating number, I first converted it to
a number with only one decimal value. So each time the user moved the controler, it would give a
decimal value like "2.12235" wich was converted to "2.1" and then it searched in an array of possible
sizes for the closest value to it, and finally assigned the size of the squares to be used.